edit reporte cierre campaña y mas cosas

+++ AgroForm.Business/Contracts/IServiceBase.cs
@@ -13,15 +13,15 @@ namespace AgroForm.Business.Contracts
         Task<OperationResult<List<T>>> GetAllAsync();
         Task<OperationResult<List<T>>> GetAllByCamapniaAsync();
         Task<OperationResult<List<T>>> GetAllWithDetailsAsync();
-        Task<OperationResult<T>> GetByIdAsync(long id);
-        Task<OperationResult<T>> GetByIdWithDetailsAsync(long id);
+        Task<OperationResult<T>> GetByIdAsync(int id);
+        Task<OperationResult<T>> GetByIdWithDetailsAsync(int id);
         Task<OperationResult<T>> CreateAsync(T entity);
         Task<OperationResult<List<T>>> CreateRangeAsync(List<T> entityList);
         Task<OperationResult<T>> UpdateAsync(T entity);
         Task<OperationResult<List<T>>> UpdateRangeAsync(List<T> entityList);
-        Task<OperationResult> DeleteAsync(long id);
-        Task<OperationResult> DeleteRangeAsync(List<long> ids);
-        Task<bool> ExistsAsync(long id);
+        Task<OperationResult> DeleteAsync(int id);
+        Task<OperationResult> DeleteRangeAsync(List<int> ids);
+        Task<bool> ExistsAsync(int id);
         Task<OperationResult> ValidateAsync(T entity);
         IQueryable<T> GetQuery();
     }
